# backend/doc/openapi.yaml
openapi: 3.1.0
info:
  title: Soniq API
  version: 0.1.0
  description: Local-first API for identity, workspaces, recording uploads, processing status, and transcript/summary results.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
    description: Same-origin local API server.
security:
  - CookieAuth: []
paths:
...
  /auth/signup:
    post:
      summary: Sign up with email and password
      operationId: signUp
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '201':
          description: User account, default workspace, and session created.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: httpOnly `soniq_session` cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /auth/signin:
    post:
      summary: Sign in with email and password
      operationId: signIn
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
      responses:
        '200':
          description: Sign-in accepted and session cookie issued.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: httpOnly `soniq_session` cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /auth/signout:
    post:
      summary: Sign out current session
      operationId: signOut
      responses:
        '204':
          description: Session revoked when present and cookie cleared.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Expired httpOnly `soniq_session` cookie.
        default:
          $ref: '#/components/responses/Error'
  /me:
    get:
      summary: Get current user
      operationId: getMe
      responses:
        '200':
          description: Current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /workspaces:
    get:
      summary: List workspaces for the current user
      operationId: listWorkspaces
      responses:
        '200':
          description: Workspaces available to the current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListWorkspacesResponse'
        '401':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /workspaces/{workspace_id}/recordings:
    get:
      summary: List workspace recordings
      operationId: listRecordings
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 100
            default: 50
      responses:
        '200':
          description: Recent recordings in the workspace.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRecordingsResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    post:
      summary: Create recording metadata
      description: Creates a recording session in a workspace without uploading audio or starting processing.
      operationId: createRecording
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecordingRequest'
      responses:
        '201':
          description: Recording created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recording'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /workspaces/{workspace_id}/recordings/upload:
    post:
      summary: Upload recording audio
      description: Uploads audio into a workspace, creates a recording, and starts processing.
      operationId: uploadRecording
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [workflow_type, audio]
              properties:
                title:
                  type: string
                workflow_type:
                  $ref: '#/components/schemas/WorkflowType'
                language:
                  type: string
                  example: zh
                audio:
                  type: string
                  format: binary
      responses:
        '201':
          description: Recording created and processing enqueue attempted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadRecordingResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /workspaces/{workspace_id}/recordings/{recording_id}:
    get:
      summary: Get recording
      operationId: getRecording
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
        - $ref: '#/components/parameters/RecordingID'
      responses:
        '200':
          description: Recording found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recording'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /workspaces/{workspace_id}/recordings/{recording_id}/status:
    get:
      summary: Get recording status
      operationId: getRecordingStatus
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
        - $ref: '#/components/parameters/RecordingID'
      responses:
        '200':
          description: Recording status found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingStatusResponse'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /workspaces/{workspace_id}/recordings/{recording_id}/retry:
    post:
      summary: Retry failed recording processing
      description: Resets a failed audio-backed recording and attempts to enqueue processing again.
      operationId: retryRecording
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
        - $ref: '#/components/parameters/RecordingID'
      responses:
        '200':
          description: Recording retry was requested.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetryRecordingResponse'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /workspaces/{workspace_id}/recordings/{recording_id}/details:
    get:
      summary: Get recording details
      description: Returns recording metadata with transcript, transcript segments, and summary when available.
      operationId: getRecordingDetails
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
        - $ref: '#/components/parameters/RecordingID'
      responses:
        '200':
          description: Recording details found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingDetails'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: soniq_session
  parameters:
    WorkspaceID:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
      description: Workspace identifier.
    RecordingID:
      name: recording_id
      in: path
      required: true
      schema:
        type: string
      description: Recording identifier.
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    AuthUserResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/User'
    SignInRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 1024
    SignUpRequest:
      type: object
      required: [email, display_name, password]
      properties:
        email:
          type: string
          format: email
        display_name:
          type: string
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 1024
    User:
      type: object
      required: [id, email, display_name, created_at, updated_at]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WorkspaceRole:
      type: string
      enum: [owner, member]
    Workspace:
      type: object
      required: [id, name, role, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        role:
          $ref: '#/components/schemas/WorkspaceRole'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ListWorkspacesResponse:
      type: object
      required: [workspaces]
      properties:
        workspaces:
          type: array
          items:
            $ref: '#/components/schemas/Workspace'
    WorkflowType:
      type: string
      enum: [meeting, lecture, interview, memo]
    RecordingStatus:
      type: string
      enum: [uploaded, processing, transcribing, summarizing, completed, failed, cancelled]
    Recording:
      type: object
      required: [id, workspace_id, title, status, workflow_type, language, created_at, updated_at]
      properties:
        id:
          type: string
        workspace_id:
          type: string
        title:
          type: string
        status:
          $ref: '#/components/schemas/RecordingStatus'
        workflow_type:
          $ref: '#/components/schemas/WorkflowType'
        language:
          type: string
        audio_object_key:
          type: string
        audio_content_type:
          type: string
        audio_size_bytes:
          type: integer
          format: int64
        failure_reason:
          type: string
        completed_at:
          type: string
          format: date-time
        failed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateRecordingRequest:
      type: object
      required: [workflow_type]
      properties:
        title:
          type: string
        workflow_type:
          $ref: '#/components/schemas/WorkflowType'
        language:
          type: string
          example: en
    ListRecordingsResponse:
      type: object
      required: [recordings]
      properties:
        recordings:
          type: array
          items:
            $ref: '#/components/schemas/Recording'
    UploadRecordingResponse:
      type: object
      required: [recording, processing_enqueued]
      properties:
        recording:
          $ref: '#/components/schemas/Recording'
        processing_enqueued:
          type: boolean
    RecordingStatusResponse:
      type: object
      required: [id, workspace_id, status]
      properties:
        id:
          type: string
        workspace_id:
          type: string
        status:
          $ref: '#/components/schemas/RecordingStatus'
        failure_reason:
          type: string
        completed_at:
          type: string
          format: date-time
        failed_at:
          type: string
          format: date-time
    RetryRecordingResponse:
      type: object
      required: [recording, processing_enqueued]
      properties:
        recording:
          $ref: '#/components/schemas/Recording'
        processing_enqueued:
          type: boolean
    RecordingTranscript:
      type: object
      required: [recording_id, provider, model, language, text, transcribed_at]
      properties:
        recording_id:
          type: string
        provider:
          type: string
        model:
          type: string
        language:
          type: string
        text:
          type: string
        transcribed_at:
          type: string
          format: date-time
    RecordingTranscriptSegment:
      type: object
      required: [id, recording_id, segment_index, start_ms, end_ms, speaker_label, text, confidence]
      properties:
        id:
          type: string
        recording_id:
          type: string
        segment_index:
          type: integer
        start_ms:
          type: integer
        end_ms:
          type: integer
        speaker_label:
          type: string
        text:
          type: string
        confidence:
          type: number
          format: double
    RecordingSummary:
      type: object
      required: [recording_id, provider, model, type, title, overview, content_markdown, summarized_at]
      properties:
        recording_id:
          type: string
        provider:
          type: string
        model:
          type: string
        type:
          $ref: '#/components/schemas/WorkflowType'
        title:
          type: string
        overview:
          type: string
        content_markdown:
          type: string
        summarized_at:
          type: string
          format: date-time
    RecordingDetails:
      type: object
      required: [recording, segments]
      properties:
        recording:
          $ref: '#/components/schemas/Recording'
        transcript:
          anyOf:
            - $ref: '#/components/schemas/RecordingTranscript'
            - type: 'null'
        segments:
          type: array
          items:
            $ref: '#/components/schemas/RecordingTranscriptSegment'
        summary:
          anyOf:
            - $ref: '#/components/schemas/RecordingSummary'
            - type: 'null'
...
    ErrorResponse:
      type: object
      required: [code, message, status]
      properties:
        code:
          type: string
          enum:
            - unauthenticated
            - invalid_credentials
            - user_already_exists
            - validation_failed
            - forbidden
            - not_found
            - method_not_allowed
            - request_too_large
            - unsupported_media_type
            - conflict
            - rate_limited
            - internal_error
            - service_unavailable
        message:
          type: string
        status:
          type: integer